requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
flake8==7.1.2
django-redis==5.4.0
pillow==11.1.0
coverage==7.8.0
drf-yasg==1.21.10
django-cors-headers==4.7.0